// src/types/slides.ts
export type ChartType =
  | 'bar'
  | 'column'
  | 'line'
  | 'pie'
  | 'ring'
  | 'area'
  | 'radar'
  | 'scatter'
export interface ChartOptions {
  lineSmooth?: boolean
  stack?: boolean
}

export interface ChartData {
  labels: string[]
  legends: string[]
  series: number[][]
}